Home dev 1

+++ Wishpang/account/models.py
@@ -7,4 +7,4 @@ from django.db import models
 class Profile(models.Model):
 	user = models.OneToOneField(User, on_delete=models.CASCADE)
 	nickname = models.TextField(max_length=50, null=True)
-	kakao_id = models.IntegerField(null=False)
+	kakao_id = models.BigIntegerField(null=False)
